feat: 整合代码
重构部分代码，`TODO：提示`为待修改内容
