+++ packages/application/dragger-editor/src/components/Right/Right.tsx
@@ -5,31 +5,84 @@ import React, { useContext } from 'react'
 // import Button from '@mui/material/Button'
 import './Right.css'
 import ProfileForm from '@cdl-pkg/profile-form'
-import { Context } from '../../store'
-const selectedNode = {
-  tag: 'MyCom',
-  properties: {
-    width: {
-      type: 'number',
-      minimum: 1,
-      value: 50,
-      maximum: 2000,
-      interaction: 'slide'
-    },
-    height: {
-      type: 'number',
-      minimum: 1,
-      value: 50,
-      maximum: 2000,
-      interaction: 'slide'
-    }
+import {selectedNode, getSchema, dispatch} from './mock'
+const schema = getSchema('1')
+export default class Right extends React.Component {
+  render(): React.ReactNode {
+    return (
+      <div className='right'>
+        {/* @ts-ignore */}
+        <ProfileForm tag="myCom" selectedNode={selectedNode} schema={schema} dispatch={dispatch}/>
+      </div>
+      // <div className='right'>
+      //   <div className='top-btn'>
+      //     <Button variant='outlined'>button</Button>
+      //   </div>
+      //   <div className='small-icon'>
+      //     <IconButtons />
+      //   </div>
+      //   <div className='props'>
+      //     <ul className='props-list'>
+      //       <li>
+      //         <div>
+      //           <div>Text</div>
+      //           <TextField
+      //             size='small'
+      //             id='outlined-basic'
+      //             variant='outlined'
+      //           />
+      //         </div>
+      //       </li>
+      //       <li>
+      //         <div>
+      //           <div>Size</div>
+      //           {/* <BasicSelect /> */}
+      //           <select name='' id=''></select>
+      //         </div>
+      //       </li>
+      //       <li>
+      //         <div>
+      //           <div>Variant</div>
+      //           {/* <BasicSelect/> */}
+      //           <select name='' id=''></select>
+      //         </div>
+      //       </li>
+      //       <li>
+      //         <div>
+      //           <div>Color Scheme</div>
+      //           <TextField
+      //             size='small'
+      //             id='outlined-basic'
+      //             variant='outlined'
+      //           />
+      //         </div>
+      //       </li>
+      //       <li>
+      //         <div>
+      //           <div>Left icon</div>
+      //           <TextField
+      //             size='small'
+      //             id='outlined-basic'
+      //             variant='outlined'
+      //           />
+      //         </div>
+      //       </li>
+      //       <li>
+      //         <div>
+      //           <div>Right icon</div>
+      //           <TextField
+      //             size='small'
+      //             id='outlined-basic'
+      //             variant='outlined'
+      //           />
+      //         </div>
+      //       </li>
+      //     </ul>
+      //   </div>
+      //   <div>
+      //     <ControlledAccordions />
+      //   </div>
+      // </div>
+    )
   }
-}
-export default function Right() {
-  const { components, dispatch } = useContext(Context)
-  return (
-    <div className='right'>
-      <ProfileForm {...selectedNode} />
-    </div>
-  )
 }
